feat: (security): logout (#251)
