--- client/src/app/core/app/app.component.ts
@@ -1,10 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
-})
-export class AppComponent {
-  title = 'client';
-}
